Init intro branch. Create SRC folder with app
