+++ src/constants/Color.ts
@@ -1,0 +1,7 @@
+export const Color = {
+    red: 'red',
+    white: '#FFFFFF',
+    orange_tan_crayola: '#D69965',
+    cream_alabaster: '#F4F1E4',
+    green_msu: '#233D3A'
+}
